# level0: pick smallest queue ordinals