On an Arrow Lake machine where there are multiple command queue groups to create queues from, picking the lowest ordinal speed up OpenVINO inference:

* `benchmark_app -d GPU.0 -hint none -nstreams 1 -nireq 1 -b 1 -t 20 -m resnet-50/resnet-50.xml`: 8 FPS -> 24 FPS.

* `benchmark_app -d GPU.0 -hint none -nstreams 2 -nireq 4 -b 1 -t 20 -m resnet-50/resnet-50.xml`: 28 FPS -> 40 FPS.

However, reportedly gains are not seen on Intel Battlemage GPUs.